requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
fake-useragent==0.1.11
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
selenium==3.141.0
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
